Move parsing code to separate web worker
